refact: separate funcs into files
