+++ srcs/execute/execute.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ktomoya <ktomoya@student.42.fr>            +#+  +:+       +#+        */
+/*   By: kudoutomoya <kudoutomoya@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 19:03:27 by kudoutomoya       #+#    #+#             */
-/*   Updated: 2023/10/13 19:40:18 by ktomoya          ###   ########.fr       */
+/*   Updated: 2023/10/14 14:20:52 by kudoutomoya      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,48 +14,10 @@
 #include "../../includes/execute.h"
 
 /*
- * 目標: envpをexecveに渡す
+ * 目標: 関数をファイルに分ける
  */
 
-void	puterr_exit(const char *input, const char *msg)
-{
-	ft_putstr_fd(input, STDERR_FILENO);
-	ft_putstr_fd(": ", STDERR_FILENO);
-	ft_putendl_fd(msg, STDERR_FILENO);
-	exit(FAILURE);
-}
-
-void	putsyserr_exit(const char *syscall_name)
-{
-	perror(syscall_name);
-	exit(FAILURE);
-}
-
-bool	is_builtin(char *cmd)
-{
-	if (ft_strcmp(cmd, "echo") == 0
-		|| ft_strcmp(cmd, "cd") == 0
-		|| ft_strcmp(cmd, "pwd") == 0
-		|| ft_strcmp(cmd, "export") == 0
-		|| ft_strcmp(cmd, "unset") == 0
-		|| ft_strcmp(cmd, "env") == 0
-		|| ft_strcmp(cmd, "exit") == 0)
-		return (true);
-	return (false);
-}
-
-bool	is_directory(char *path)
-{
-	struct stat	buf;
-
-	if (stat(path, &buf) == ERROR)
-		putsyserr_exit("stat");
-	if (S_ISDIR(buf.st_mode))
-		return (true);
-	return (false);
-}
-
-int	execute_builtin(char *cmds[], t_env *env)
+static int	execute_builtin(char *cmds[], t_env *env)
 {
 	if (ft_strcmp(cmds[0], "echo") == 0)
 		return (builtin_echo(cmds));
@@ -73,60 +35,7 @@ int	execute_builtin(char *cmds[], t_env *env)
 		return (builtin_exit(cmds));
 }
 
-void	execute_path(char *const argv[], t_env *env)
-{
-	if (access(argv[0], X_OK) == ERROR)
-		puterr_exit(argv[0], strerror(errno));
-	if (execve(argv[0], argv, env->envp) == ERROR)
-		putsyserr_exit("execve");
-}
-
-void	execute_abspath(char *const argv[], t_env *env)
-{
-	if (access(argv[0], X_OK) == ERROR)
-		puterr_exit(argv[0], strerror(errno));
-	if (is_directory(argv[0]))
-		puterr_exit(argv[0], "is a directory");
-	if (execve(argv[0], argv, env->envp) == ERROR)
-		putsyserr_exit("execve");
-}
-
-char	*get_path_value(char *const argv[], t_env *env)
-{
-	char	*value;
-
-	value = search_env("PATH", env);
-	if (value == NULL)
-		puterr_exit(argv[0], "command not found");
-	return (value);
-}
-
-void	search_path(char *const argv[], t_env *env)
-{
-	char	*dir_set;
-	char	file_path[PATH_MAX];
-	char	*dir;
-
-	dir_set = get_path_value(argv, env);
-	dir = ft_strtok(dir_set, ":");
-	while (dir)
-	{
-		ft_strlcpy(file_path, dir, PATH_MAX);
-		ft_strlcat(file_path, "/", PATH_MAX);
-		ft_strlcat(file_path, argv[0], PATH_MAX);
-		if (access(file_path, X_OK) == 0)
-		{
-			if (execve(file_path, argv, env->envp) == ERROR)
-				puterr_exit(file_path, strerror(errno));
-		}
-		else if (errno != ENOENT)
-			puterr_exit(argv[0], strerror(errno));
-		dir = ft_strtok(NULL, ":");
-	}
-	puterr_exit(argv[0], "command not found");
-}
-
-int	execute_executable(char *const argv[], t_env *env)
+static int	execute_executable(char *const argv[], t_env *env)
 {
 	pid_t	pid;
 	int		status;
@@ -149,49 +58,12 @@ int	execute_executable(char *const argv[], t_env *env)
 	return (0);
 }
 
-int	execute_command(char *const argv[], t_env *env)
+int	execute(char *const argv[], t_env *env)
 {
 	if (is_builtin(argv[0]))
 		return (execute_builtin((char **)argv, env));
 	else
 		return (execute_executable(argv, env));
-}
-
-size_t	count_token(t_token *tok)
-{
-	t_token	*cur;
-	size_t	count;
-
-	cur = tok;
-	count = 0;
-	while (cur->type != TYPE_EOF)
-	{
-		count++;
-		cur = cur->next;
-	}
-	return (count);
-}
-
-char	**malloc_token(t_token *tok)
-{
-	t_token	*cur;
-	char	**argv;
-	size_t	count;
-	size_t	i;
-
-	cur = tok;
-	count = count_token(tok);
-	argv = ft_calloc(count + 1, sizeof(char *));
-	i = 0;
-	while (i < count)
-	{
-		argv[i] = ft_substr(cur->str, 0, cur->len);
-		if (argv[i] == NULL)
-			return (NULL);
-		cur = cur->next;
-		i++;
-	}
-	return (argv);
 }
 
 int	main(int argc, char **argv, char **envp)
@@ -217,7 +89,7 @@ int	main(int argc, char **argv, char **envp)
 		env.envp = env_to_envp(&env);
 		token = lexer(line);
 		args = malloc_token(token);
-		execute_command(args, &env);
+		execute(args, &env);
 		free_env_to_envp(env.envp);
 		free((void *)line);
 	}
