+++ srcs/execute/execute_error.c
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   execute_error.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kudoutomoya <kudoutomoya@student.42.fr>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/14 13:52:35 by kudoutomoya       #+#    #+#             */
+/*   Updated: 2023/10/14 14:02:56 by kudoutomoya      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/libft.h"
+
+void	puterr_exit(const char *input, const char *msg)
+{
+	ft_putstr_fd(input, STDERR_FILENO);
+	ft_putstr_fd(": ", STDERR_FILENO);
+	ft_putendl_fd(msg, STDERR_FILENO);
+	exit(FAILURE);
+}
+
+void	putsyserr_exit(const char *syscall_name)
+{
+	perror(syscall_name);
+	exit(FAILURE);
+}
